Mapping Refactor:
- AutoMapper REMOVED
- Expression mapping REMOVED
- IQueryableEntity/QueryableEntity removed
- Registrations refactor
- Mapping refactor: Manually

Others refactors:
- CrudApis refactor
- EquivalencesController (and down) refactor

+++ Backend/Services/Measurement/Skyrmium.Equivalent.Measurement.Domain.Entities/OrderedMeasureEquivalence.cs
@@ -5,7 +5,7 @@ namespace Skyrmium.Equivalent.Measurement.Domain.Entities
 {
    public class OrderedMeasureEquivalence : EntityBase
    {
-      public OrderedMeasureEquivalence(int id, Guid distributedId, int order, MeasureEquivalence measureEquivalence)
+      public OrderedMeasureEquivalence(long id, Guid distributedId, int order, MeasureEquivalence measureEquivalence)
          : base(id, distributedId)
       {
          this.Order = order;
